+++ src/main/java/dev/willyelton/origins/datagen/OriginsOfLifeEntityLoot.java
@@ -1,0 +1,47 @@
+package dev.willyelton.origins.datagen;
+
+import dev.willyelton.origins.OriginsOfLife;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.data.loot.packs.VanillaEntityLoot;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.level.storage.loot.LootPool;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.entries.LootItem;
+import net.minecraft.world.level.storage.loot.functions.EnchantedCountIncreaseFunction;
+import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
+import net.minecraft.world.level.storage.loot.functions.SmeltItemFunction;
+import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
+
+import java.util.Map;
+import java.util.stream.Stream;
+
+public class OriginsOfLifeEntityLoot extends VanillaEntityLoot {
+    public OriginsOfLifeEntityLoot(HolderLookup.Provider registries) {
+        super(registries);
+    }
+
+    @Override
+    public void generate() {
+        add(OriginsOfLife.AQUATIC_CREATURE.get(),
+                LootTable.lootTable()
+                        .withPool(
+                                LootPool.lootPool()
+                                        .setRolls(ConstantValue.exactly(1.0F))
+                                        .add(
+                                                LootItem.lootTableItem(OriginsOfLife.RAW_MEAT.get())
+                                                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(1.0F, 3.0F)))
+                                                        .apply(SmeltItemFunction.smelted().when(this.shouldSmeltLoot()))
+                                                        .apply(EnchantedCountIncreaseFunction.lootingMultiplier(this.registries, UniformGenerator.between(0.0F, 1.0F)))
+                                        )
+                        ));
+    }
+
+    @Override
+    protected Stream<EntityType<?>> getKnownEntityTypes() {
+        return BuiltInRegistries.ENTITY_TYPE.entrySet().stream()
+                .filter(e -> e.getKey().identifier().getNamespace().equals(OriginsOfLife.MODID))
+                .map(Map.Entry::getValue);
+    }
+}