+++ src/main/java/dev/willyelton/origins/datagen/OriginsOfLifeRecipes.java
@@ -8,12 +8,18 @@ import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeOutput;
 import net.minecraft.data.recipes.RecipeProvider;
+import net.minecraft.data.recipes.SimpleCookingRecipeBuilder;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.CampfireCookingRecipe;
+import net.minecraft.world.item.crafting.CookingBookCategory;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.SmokingRecipe;
 
 import java.util.concurrent.CompletableFuture;
 
 public class OriginsOfLifeRecipes extends RecipeProvider {
     public static final Criterion<InventoryChangeTrigger.TriggerInstance> HAS_FOSSIL = InventoryChangeTrigger.TriggerInstance.hasItems(OriginsOfLife.FOSSIL.get());
+    public static final Criterion<InventoryChangeTrigger.TriggerInstance> HAS_MEAT = InventoryChangeTrigger.TriggerInstance.hasItems(OriginsOfLife.RAW_MEAT.get());
 
     protected OriginsOfLifeRecipes(HolderLookup.Provider registries, RecipeOutput output) {
         super(registries, output);
@@ -37,6 +43,12 @@ public class OriginsOfLifeRecipes extends RecipeProvider {
                 .define('c', OriginsOfLife.CAGE)
                 .unlockedBy("has_fossil", HAS_FOSSIL)
                 .save(output);
+
+        SimpleCookingRecipeBuilder.smelting(Ingredient.of(OriginsOfLife.RAW_MEAT), RecipeCategory.FOOD, CookingBookCategory.FOOD, OriginsOfLife.COOKED_MEAT, 0.35F, 200)
+                .unlockedBy("has_mystery_meat", HAS_MEAT)
+                .save(this.output);
+        simpleCookingRecipe("smoking", SmokingRecipe::new, 100, OriginsOfLife.RAW_MEAT, OriginsOfLife.COOKED_MEAT, 0.35F);
+        simpleCookingRecipe("campfire_cooking", CampfireCookingRecipe::new, 600, OriginsOfLife.RAW_MEAT, OriginsOfLife.COOKED_MEAT, 0.35F);
     }
 
     public static class Runner extends RecipeProvider.Runner {