+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Origins of Life Changelog
+## 26.2.0.4-alpha
+- Adds Mystery Meat
+  - Eating raw meat will apply poison and hunger
+  - Can be cooked
+
 ## 26.2.0.3-alpha
-- Added the Display Case!
+- Adds the Display Case!
   - Allows you to display your creatures in a nice little glass box
   - Right-click a filled cage to insert an entity
   - Right-click with an empty cage to retrieve it
   - If you break the Display Case, the entity will spawn where you broke it
-  
+- Fixes some misaligned hitboxes
+
 ## 26.2.0.2-alpha
 - Fossil blocks now mine correctly
   - Pickaxe for deepslate and sulfur